Sort segments on flush
This change sorts the segments on flush if an index sorting is provided.
The sorting is done in the default indexing chain and reuse SortingLeafReader when possible.
Term vectors and stored fields are sorted in two steps, the first step writes the documents unsorted and the second pass sorts the output of the first phase. Each step uses the writer/reader defined by the main codec.